질문 추천 기능 추가

+++ src/main/java/com/example/Sbb/Sbb/answer/AnswerService.java
@@ -8,7 +8,7 @@ public interface AnswerService {
     public void delete(AnswerDTO answerDTO);
     public void modify(AnswerDTO answerDTO, String content);
     public void recommend(AnswerDTO answerDTO, SiteUserDTO siteUserDTO);
-    public AnswerDTO getAnswer(Integer id);
+    public AnswerDTO getAnswer(Long id);
     public AnswerDTO toDTO(AnswerEntity answerEntity);
     public AnswerEntity toEntity(AnswerDTO answerDTO);
 }
